Change macppc's ci_idepth from -1 to 0 when no interrupts
Now curcpu()->ci_idepth == 0 would work on macppc as it does on
powerpc64 and other archs, by checking that we aren't in an interrupt.
This also makes macppc's splassert_check() less different.

locore.S checks for the 1st interrupt when it increments ci_idepth.
Using -1 for no interrupts had saved an instruction, as "addic." was
short for "addi; cmpwi".  Using 0, it now does "cmpwi; addi" to
compare ci_idepth with 0 before incrementing it.

ok mpi@