In powerpc64 pmap, when replacing a pte, put it in the correct pteg
In pte_insert(), if both the primary page table entry group and the
secondary pteg are full, then we delete an old pte and replace it with
the new pte.  We might have set "idx" wrong and inserted the new pte
into the wrong pteg (with the wrong PTE_HID bit).  This problem almost
never happened; it is rare for both ptegs to be full.

When the loop "for (try = 0; try < 16; try++)" looks for a slot for
the new pte, set "idx" to the new pte's primary pteg.  Put it in the
primary pteg "idx" or secondary pteg "idx ^ pmap_ptab_mask".  When we
delete the old pte, set "idx" to the old pte's primary pteg.

Eric Gosse reported a bug where the kernel crashed on an unexpected
fault.  I suspect that a pte fell into the wrong "idx" and got lost.

ok kettenis@