sys/ihidev: quirk for Honor MagicBook's touchpad
Without this quirk a Honor MagicBook's touchpad can not be read as
multioutch. This touchpad needs re-power command with sleep after report
description is read. Without waitng quite enough, almost the seconds, or
re-power command it reads only as:

	imt0 at ihidev1ihidev1: failed fetching report
	ims0 at ihidev1 reportid 1: 3 buttons, Z dir

but seems only on the cold boot. Otherwise with 1 seconds sleep and re-power
command, or on warm boot, it reads as:

	imt0 at ihidev1: touchpad, 5 contacts
	ims0 at ihidev1 reportid 1: 3 buttons, Z dir

OK: mglocker@ tobhe@