Lock the new socket in sonewconn() also for the internet family.
UNIX sockets were already locking the newly allocated socket during
soreserve() and pru_attach().  The same is neeed for TCP input
unlocking.  When two sockets have to be locked, one needs shared
net lock, but the other already has it.  This is the same idea as
in doaccept().  Unify the lock functions solock_nonet() and
sounlock_nonet().

OK mvs@