Toeplitz hash for UDP and IPv6 TCP output.
IPv4 TCP output uses the toeplitz hash as flow id.  It is calculated
in in_pcbconnect() and can be used for all connected sockets.  Add
it for UDP and TCP IPv6, too.  As pf calculates its own hash, this
affects only setups with pf disabled.  It gives an improvement in
traffic distribution over the queues and 20% performance increase
with UDP send on v4/v6 and TCP send on v6 without pf.

tested and OK sf@