Move ipsec-enc-alg, ipsec-auth-alg and ipsec-comp-alg sysctl variables out of netlock.
Use integers instead of strings for in-kernel representation. So we can
update them atomically. Also we remove strcmp() from the hot (?) path.

Note, behavior was changed for a little. Against previous it denies to
set incorrect values. In such case the EINVAL error will be returned to
userland.

ok bluhm