add support for the ERSPAN Type II protocol
ERSPAN is a specific GRE 0 protocol id with GRE sequence numbers
enabled, with it's own shim header, and then an Ethernet payload.

it is intended for use as a network based packet capture format,
but it's implemented as an Ethernet tunnel interface here. as an
interface, it reuses all the BPF based infrastructure and tooling
available to us. it's able to send packets too, which allows it to
be added as a span port to bridge/veb setups.

the driver can also act as a collector for any ERSPAN Type II packets
sent to it. it's also set up with an extra BPF DLT_LOOP interface
that includes the encapsulation headers so you can tell where the
encapsulated packets came from.

ok claudio@ sthen@