stop defining CONFIG_FRAMEBUFFER_CONSOLE
If defined, initial_plane_vma() won't use the framebuffer setup by bios
when it is over half of the stolen size.

yasuoka@ tracked down an inteldrm problem with GuC failing to initialise
on hibernate resume to this change.  Specifically on a VAIO VJPK23 where
the 3840x2160 framebuffer is over half of the stolen size.