fix deadline calculation against what timeouts actually run.
deadlines are added to the current ticks value, so if you want to
wait one tick you add a 1 to ticks. however, when the soft interrupt
processes the lists of pending timeouts, ticks gets incremented
before all the comparisons are made. this means the timeout you
wanted to run after 1 tick in the future actually runs in the very
next tick, which is too short.

anton found that the sem_timedwait regress tripped over this.

tested by anton@
ok claudio@ anton@