have route sourceaddr use RTF_GATEWAY to decide when to kick in.
previously it used !RTF_HOST and !RTF_LLINFO.

the intention with route sourceaddr was to use it except when a
peer was on link. however, it is possible to have host routes (ie,
RTF_HOST) via a gateway, which ended up not using the route sourceaddr
when it should not have. by definition any route with RTF_GATEWAY
set is not directly connected, so using it seems to better suit
what route sourceaddr is doing.

discussed with and ok claudio@ denis@
tested by denis@