add an API for cpu xcalls (crosscalls)
this lets you dispatch code to run on the specified cpu from a intr
(IPL_SOFTCLOCK softintr) context. this is less overhead and mess
than pegging the current proc to the relevant cpu.

this requires a bit of MD glue to wire it into each architecture,
and can be done as needed.

ok kettenis@