Unlock IPV6CTL_MAXFRAGS case of ip6_sysctl().
The frag6_input() assumes that `ip6_maxfrags' could be negative and
accept all the fragments without the limit in this case, meanwhile the
sysctl(2) interface denies negative values for `ip6_maxfrags'. Unlimited
queue is bad thing, so adjust the frag6_input() to be consistent with
sysctl(2). Don't check `ip6_maxfrags' against null because in such case
the "frag6_nfrags >= ip6_maxfrags" condition is always true. Unused in
ramdisk.

ok bluhm