Handle #VC in locore0 due to dr[67] write intercepts
When running on Linux/KVM with SEV-ES enabled, writes to dr[67]
raise a #VC trap.  We have no means to emulate writes to dr[67]
unsing the MSR protocol.  Thus skip these instructions, when #VC
is raised in locore0.

On vmm(4)/vmd(8) we are runnign in compat mode and writes to dr[67]
are not intercepted; thus only adjust the 32-bit case.

ok sf mlarkin